+++ src/pynwb/ecephys.py
@@ -231,7 +231,9 @@ class Clustering(NWBDataInterface):
              'shape': (None,)},
             {'name': 'name', 'type': str, 'doc': 'the name of this container', 'default': 'Clustering'})
     def __init__(self, **kwargs):
-        warnings.warn("use pynwb.misc.Units or NWBFile.units instead", DeprecationWarning)
+        self._error_on_new_pass_on_construct(
+            error_msg='The Clustering neurodata type is deprecated. Use pynwb.misc.Units or NWBFile.units instead'
+        )
         args_to_set = popargs_to_dict(('description', 'num', 'peak_over_rms', 'times'), kwargs)
         super().__init__(**kwargs)
         args_to_set['peak_over_rms'] = list(args_to_set['peak_over_rms'])
@@ -265,7 +267,9 @@ class ClusterWaveforms(NWBDataInterface):
              'doc': 'the standard deviations of waveforms for each cluster'},
             {'name': 'name', 'type': str, 'doc': 'the name of this container', 'default': 'ClusterWaveforms'})
     def __init__(self, **kwargs):
-        warnings.warn("use pynwb.misc.Units or NWBFile.units instead", DeprecationWarning)
+        self._error_on_new_pass_on_construct(
+            error_msg='The ClusterWaveforms neurodata type is deprecated. Use pynwb.misc.Units or NWBFile.units instead'
+        )
         args_to_set = popargs_to_dict(('clustering_interface', 'waveform_filtering',
                                        'waveform_mean', 'waveform_sd'), kwargs)
         super().__init__(**kwargs)