+++ src/pynwb/icephys.py
@@ -345,9 +345,11 @@ class SweepTable(DynamicTable):
              'default': "A sweep table groups different PatchClampSeries together."},
             *get_docval(DynamicTable.__init__, 'id', 'columns', 'colnames'))
     def __init__(self, **kwargs):
-        warnings.warn("Use of SweepTable is deprecated. Use the IntracellularRecordingsTable "
-                      "instead. See also the  NWBFile.add_intracellular_recordings function.",
-                      DeprecationWarning)
+        error_msg=('SweepTable is deprecated. Use the IntracellularRecordingsTable instead. '
+                   'See also the NWBFile.add_intracellular_recordings function.')
+        if not self._in_construct_mode:
+            raise ValueError(error_msg)
+
         super().__init__(**kwargs)
 
     @docval({'name': 'pcs', 'type': PatchClampSeries,