+++ tests/unit/test_icephys.py
@@ -39,26 +39,31 @@ class NWBFileICEphys(TestCase):
     def setUp(self):
         self.icephys_electrode = GetElectrode()
 
-    def test_sweep_table_depractation_warn(self):
-        msg = ("Use of SweepTable is deprecated. Use the IntracellularRecordingsTable "
-               "instead. See also the  NWBFile.add_intracellular_recordings function.")
-        with self.assertWarnsWith(DeprecationWarning, msg):
-            _ = NWBFile(
-                session_description='NWBFile icephys test',
-                identifier='NWB123',  # required
-                session_start_time=datetime(2017, 4, 3, 11, tzinfo=tzlocal()),
-                ic_electrodes=[self.icephys_electrode, ],
-                sweep_table=SweepTable())
+    def test_sweep_table_deprecation_warn(self):
+        msg = ("SweepTable is deprecated. Use the IntracellularRecordingsTable instead. "
+               "See also the NWBFile.add_intracellular_recordings function.")
+        
+        with self.assertRaisesWith(ValueError, msg):
+            SweepTable()
 
-    def test_ic_electrodes_parameter_deprecation(self):
-        # Make sure we warn when using the ic_electrodes parameter on NWBFile
-        msg = "Use of the ic_electrodes parameter is deprecated. Use the icephys_electrodes parameter instead"
-        with self.assertWarnsWith(DeprecationWarning, msg):
-            _ = NWBFile(
-                session_description='NWBFile icephys test',
+        # create object in construct mode, modeling the behavior of the ObjectMapper on read
+        # should not raise error or warning
+        sweepT = SweepTable.__new__(SweepTable, in_construct_mode=True)
+        sweepT.__init__()
+        
+        kwargs = dict(session_description='NWBFile icephys test',
                 identifier='NWB123',  # required
                 session_start_time=datetime(2017, 4, 3, 11, tzinfo=tzlocal()),
-                ic_electrodes=[self.icephys_electrode, ])
+                icephys_electrodes=[self.icephys_electrode, ],
+                sweep_table=sweepT)
+        
+        with self.assertRaisesWith(ValueError, msg):
+            NWBFile(**kwargs)
+
+        # create object in construct mode, modeling the behavior of the ObjectMapper on read
+        # should not warn or error
+        nwbfile = NWBFile.__new__(NWBFile, in_construct_mode=True)
+        nwbfile.__init__(**kwargs)
 
     def test_icephys_electrodes_parameter(self):
         nwbfile = NWBFile(
@@ -68,52 +73,17 @@ class NWBFileICEphys(TestCase):
                 icephys_electrodes=[self.icephys_electrode, ])
         self.assertEqual(nwbfile.get_icephys_electrode('test_iS'), self.icephys_electrode)
 
-    def test_add_ic_electrode_deprecation(self):
-        # Make sure we warn when using the add_ic_electrodes parameter on NWBFile
-        nwbfile = NWBFile(
-                session_description='NWBFile icephys test',
-                identifier='NWB123',  # required
-                session_start_time=datetime(2017, 4, 3, 11, tzinfo=tzlocal()))
-
-        msg = "NWBFile.add_ic_electrode has been replaced by NWBFile.add_icephys_electrode."
-        with self.assertWarnsWith(DeprecationWarning, msg):
-            nwbfile.add_ic_electrode(self.icephys_electrode)
-
     def test_ic_electrodes_attribute_deprecation(self):
         nwbfile = NWBFile(
             session_description='NWBFile icephys test',
             identifier='NWB123',  # required
             session_start_time=datetime(2017, 4, 3, 11, tzinfo=tzlocal()),
             icephys_electrodes=[self.icephys_electrode, ])
-        # make sure NWBFile.ic_electrodes property warns
-
-        msg = "NWBFile.ic_electrodes has been replaced by NWBFile.icephys_electrodes."
-        with self.assertWarnsWith(DeprecationWarning, msg):
-            nwbfile.ic_electrodes
 
         # make sure NWBFile.get_ic_electrode warns
-        msg = "NWBFile.get_ic_electrode has been replaced by NWBFile.get_icephys_electrode."
-        with self.assertWarnsWith(DeprecationWarning, msg):
+        msg = "'NWBFile' object has no attribute 'get_ic_electrode'"
+        with self.assertRaisesWith(AttributeError, msg):
             nwbfile.get_ic_electrode(self.icephys_electrode.name)
-
-    def test_create_ic_electrode_deprecation(self):
-        nwbfile = NWBFile(
-            session_description='NWBFile icephys test',
-            identifier='NWB123',  # required
-            session_start_time=datetime(2017, 4, 3, 11, tzinfo=tzlocal()))
-        device = Device(name='device_name')
-        msg = "NWBFile.create_ic_electrode has been replaced by NWBFile.create_icephys_electrode."
-        with self.assertWarnsWith(DeprecationWarning, msg):
-            nwbfile.create_ic_electrode(
-                name='test_iS',
-                device=device,
-                description='description',
-                slice='slice',
-                seal='seal',
-                location='location',
-                resistance='resistance',
-                filtering='filtering',
-                initial_access_resistance='initial_access_resistance')
 
 
 class IntracellularElectrodeConstructor(TestCase):