+++ tests/unit/test_image.py
@@ -154,7 +154,7 @@ class ImageSeriesConstructor(TestCase):
                     "ImageSeries 'test_iS': The number of frame indices in "
                     "'starting_frame' should have the same length as 'external_file'."
                 )
-                # Create the image series in construct mode, modelling the behavior
+                # Create the image series in construct mode, modeling the behavior
                 # of the ObjectMapper on read while avoiding having to create, write,
                 # and read and entire NWB file
                 obj = ImageSeries.__new__(ImageSeries,
@@ -243,17 +243,23 @@ class ImageSeriesConstructor(TestCase):
         """Test that format is set to 'external' if not provided, when external_file is provided."""
         msg = (
             "ImageSeries 'test_iS': The value for 'format' has been changed to 'external'. "
-            "Setting a default value for 'format' is deprecated and will be changed to "
-            "raising a ValueError in the next major release."
+            "If an external file is detected, setting a value for "
+            "'format' other than 'external' is deprecated."
         )
-        with self.assertWarnsWith(DeprecationWarning, msg):
-            iS = ImageSeries(
-                name="test_iS",
+        kwargs = dict(name="test_iS",
                 external_file=["external_file", "external_file2"],
                 unit="n.a.",
                 starting_frame=[0, 10],
-                rate=0.2,
-            )
+                rate=0.2,)
+        
+        # create object with deprecated argument
+        with self.assertRaisesWith(ValueError, msg):
+            ImageSeries(**kwargs)
+
+        # create object in construct mode, modeling the behavior of the ObjectMapper on read
+        iS = ImageSeries.__new__(ImageSeries, in_construct_mode=True)
+        with self.assertWarnsWith(warn_type=UserWarning, exc_msg=msg):
+            iS.__init__(**kwargs)
         self.assertEqual(iS.format, "external")
 
     def test_external_file_with_correct_format(self):
@@ -309,6 +315,23 @@ class ImageSeriesConstructor(TestCase):
                 rate=0.2,
             )
 
+    def test_bits_per_pixel_deprecation(self):
+        msg = "bits_per_pixel is deprecated"
+        kwargs = dict(name='test_iS',
+                unit='unit',
+                external_file=['external_file'],
+                starting_frame=[0],
+                format='external',
+                timestamps=[1., 2., 3.],
+                bits_per_pixel=8)
+        with self.assertRaisesWith(ValueError, msg):
+            ImageSeries(**kwargs)
+
+        # create object in construct mode, modeling the behavior of the ObjectMapper on read
+        # no warning or error should be raised
+        iS = ImageSeries.__new__(ImageSeries, in_construct_mode=True)
+        iS.__init__(**kwargs)
+
 
 class IndexSeriesConstructor(TestCase):
 
@@ -329,23 +352,29 @@ class IndexSeriesConstructor(TestCase):
         self.assertEqual(iS.unit, 'N/A')
         self.assertIs(iS.indexed_images, images)
 
-    def test_init_bad_unit(self):
+    def test_init_bad_unit(self):        
         ts = TimeSeries(
             name='test_ts',
             data=[1, 2, 3],
             unit='unit',
             timestamps=[0.1, 0.2, 0.3]
         )
-        msg = ("The 'unit' field of IndexSeries is fixed to the value 'N/A' starting in NWB 2.5. Passing "
-               "a different value for 'unit' will raise an error in PyNWB 3.0.")
-        with self.assertWarnsWith(PendingDeprecationWarning, msg):
-            iS = IndexSeries(
-                name='test_iS',
+        msg = ("The 'unit' field of IndexSeries is fixed to the value 'N/A'.")
+        kwargs = dict(name='test_iS',
                 data=[1, 2, 3],
                 unit='na',
                 indexed_timeseries=ts,
-                timestamps=[0.1, 0.2, 0.3]
-            )
+                timestamps=[0.1, 0.2, 0.3])
+
+        # create object with deprecated argument
+        with self.assertRaisesWith(ValueError, msg):
+            IndexSeries(**kwargs)
+
+        # create object in construct mode, modeling the behavior of the ObjectMapper on read
+        # no warning or error should be raised
+        iS = IndexSeries.__new__(IndexSeries, in_construct_mode=True)
+        iS.__init__(**kwargs)
+
         self.assertEqual(iS.unit, 'N/A')
 
     def test_init_indexed_ts(self):
@@ -355,17 +384,30 @@ class IndexSeriesConstructor(TestCase):
             unit='unit',
             timestamps=[0.1, 0.2, 0.3]
         )
-        msg = ("The indexed_timeseries field of IndexSeries is discouraged and will be deprecated in "
-               "a future version of NWB. Use the indexed_images field instead.")
-        with self.assertWarnsWith(PendingDeprecationWarning, msg):
-            iS = IndexSeries(
-                name='test_iS',
+        msg = ("The indexed_timeseries field of IndexSeries is deprecated. "
+               "Use the indexed_images field instead.")
+        kwargs = dict(name='test_iS',
                 data=[1, 2, 3],
                 unit='N/A',
                 indexed_timeseries=ts,
-                timestamps=[0.1, 0.2, 0.3]
-            )
+                timestamps=[0.1, 0.2, 0.3])
+        
+        # create object with deprecated argument
+        with self.assertRaisesWith(ValueError, msg):
+            IndexSeries(**kwargs)
+
+        # create object in construct mode, modeling the behavior of the ObjectMapper on read
+        # no warning or error should be raised
+        iS = IndexSeries.__new__(IndexSeries, in_construct_mode=True)
+        iS.__init__(**kwargs)
+
         self.assertIs(iS.indexed_timeseries, ts)
+
+    def test_init_no_indexed_ts_or_timeseries(self):
+        msg = ("Either indexed_timeseries or indexed_images "
+               "must be provided when creating an IndexSeries.")
+        with self.assertRaisesWith(ValueError, msg):
+            IndexSeries(name='test_iS', data=[1, 2, 3], unit='N/A', timestamps=[0.1, 0.2, 0.3])
 
 
 class ImageMaskSeriesConstructor(TestCase):